mcp-server/requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
click==8.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.17.0
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.37.0
rpds-py==0.27.1
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
psycopg2-binary==2.9.10